requirements.txt
pandas==2.2.3
curl-cffi==0.7.4
beautifulsoup4==4.12.3
gspread==6.1.4
playwright==1.49.1
flake8==7.1.1
pytest==8.3.4
setuptools==75.6.0
python-telegram-bot==21.9
python-dotenv==1.0.1
playwright-stealth==1.0.6
fake-useragent==2.0.3